pyproject.toml
[project]
name        = "microservice-instant-messages"
version     = "0.1.0"
description = "Production-grade FastAPI microservice for sending instant messages (Microsoft Teams via webhooks; extensible to other channels)."
readme      = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    # kept so the preserved artifacts/main.py script still runs:
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode     = "auto"
testpaths        = ["tests"]
filterwarnings   = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]   # long lines allowed (we deliberately align columns)
